refactor: move all of the file conversion into folder
