--- src/lib/parser/experimental/VertexAPI/removeFirstAndLastLine.ts
@@ -1,11 +1,0 @@
-/**
- * Google VertexAI is returning Markdown:
- * ```html
- * [...]
- * ```
- * So we need to remove the first and last line
- */
-export function removeFirstAndLastLine(content: string): string {
-  const lines = content.split('\n');
-  return lines.slice(1, -1).join('\n');
-}
